24.08.2021-rev01 Изменил структуру проекта. Перенес Dto-модели в Web проект. Изменил маппинг данных
