+++ BPMSystem.BLL/Interfaces/IDepartmentService.cs
@@ -1,4 +1,4 @@
-﻿using BPMSystem.BLL.DTO;
+﻿using BPMSystem.DAL.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,10 +9,10 @@ namespace BPMSystem.BLL.Interfaces
 {
     public interface IDepartmentService
     {
-        Task<IEnumerable<DtoDepartment>> GetAllDepartment();
-        Task<DtoDepartment> GetDepartment(Guid id);
-        Task CreateDepartment(DtoCreateDepartment dtoDepartment);
-        Task UpdateDepartment(DtoDepartment dtoDepartment);
+        Task<IEnumerable<Department>> GetAllDepartment();
+        Task<Department> GetDepartment(Guid id);
+        Task CreateDepartment(Department dtoDepartment);
+        Task UpdateDepartment(Department dtoDepartment);
         Task DeleteDepartment(Guid id);
     }
 }
