+++ BPMSystem.BLL/Interfaces/IEmployeeService.cs
@@ -1,6 +1,4 @@
-﻿using BPMSystem.BLL.DTO.Employee;
-using BPMSystem.BLL.DTO.Employees;
-using BPMSystem.DAL.Entities;
+﻿using BPMSystem.DAL.Entities;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -9,10 +7,10 @@ namespace BPMSystem.BLL.Interfaces
 {
     public interface IEmployeeService
     {
-        Task<IEnumerable<DtoEmployee>> GetAllEmployee();
-        Task<DtoEmployee> GetEmployee(Guid id);
-        Task CreateEmployee(DtoCreateEmployee dtoEmployee);
-        Task UpdateEmployee(DtoEmployee dtoEmployee);
+        Task<IEnumerable<Employee>> GetAllEmployee();
+        Task<Employee> GetEmployee(Guid id);
+        Task CreateEmployee(Employee dtoEmployee);
+        Task UpdateEmployee(Employee dtoEmployee);
         Task DeleteEmployee(Guid id);
     }
 }
